refactor: dot-scope the bookmarklet-owned lib files
`bookmarklet.` prefix on everything the feature owns, tests included, so
dropping the bookmarklet one day is a matter of deleting its files rather
than working out which of lib/ belonged to it. The shadcn components stay
as they are — they are generated files that happen to read the portal
container context.
